Modularize simulator-specific DV flows
