+++ verilog/private/BUILD
@@ -4,5 +4,8 @@ package(default_visibility = ["//visibility:public"])
 exports_files([
     "rtl.bzl",
     "dv.bzl",
+    "simulators/pldm.bzl",
+    "simulators/vcs.bzl",
+    "simulators/xcelium.bzl",
     "verilog.bzl",
 ])
